feat: 패키지 구조 변경 및 파일 위치 이동 #6
