--- src/main/java/com/semtleapp/semtleapp/common/file/FileDownloadException.java
@@ -1,11 +1,0 @@
-package com.semtleapp.semtleapp.common.file;
-
-public class FileDownloadException extends RuntimeException{
-    public FileDownloadException(String message) {
-        super(message);
-    }
-
-    public FileDownloadException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
