--- src/main/java/com/semtleapp/semtleapp/common/file/FileUploadException.java
@@ -1,12 +1,0 @@
-package com.semtleapp.semtleapp.common.file;
-
-public class FileUploadException extends RuntimeException {
-    public FileUploadException(String message) {
-        super(message);
-    }
-
-    public FileUploadException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
-
