--- src/main/java/com/semtleapp/semtleapp/common/file/FileUploadProperties.java
@@ -1,20 +1,0 @@
-package com.semtleapp.semtleapp.common.file;
-
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@ConfigurationProperties(prefix = "file")
-@Component
-public class FileUploadProperties {
-    //@Value("${file.upload-dir}")
-    private String uploadDir;
-
-    public String getUploadDir() {
-        return uploadDir;
-    }
-
-    public void setUploadDir(String uploadDir) {
-        this.uploadDir = uploadDir;
-    }
-}
