--- src/main/java/com/semtleapp/semtleapp/common/service/FileService.java
@@ -1,16 +1,0 @@
-package com.semtleapp.semtleapp.common.service;
-
-import org.springframework.core.io.Resource;
-import org.springframework.http.ResponseEntity;
-
-import javax.servlet.http.HttpServletRequest;
-import java.util.Map;
-
-public interface FileService {
-
-    ResponseEntity<Resource> loadFileAsResource(HttpServletRequest request, String fileName);
-
-    ResponseEntity<Resource> showImage(Map<String, String> param);
-
-
-}
