--- src/main/java/com/semtleapp/semtleapp/common/swagger/SwaggerConfig.java
@@ -1,69 +1,0 @@
-package com.semtleapp.semtleapp.common.swagger;
-
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.bind.annotation.RestController;
-import springfox.documentation.builders.ApiInfoBuilder;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.service.ApiInfo;
-import springfox.documentation.service.ApiKey;
-import springfox.documentation.service.AuthorizationScope;
-import springfox.documentation.service.SecurityReference;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spi.service.contexts.SecurityContext;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-import java.util.Arrays;
-import java.util.List;
-
-@Configuration
-@EnableSwagger2
-public class SwaggerConfig {
-
-    public static final String AUTHORIZATION_HEADER = "Authorization";
-
-    @Bean
-    public Docket api() {
-        return new Docket(DocumentationType.OAS_30)
-                .securityContexts(Arrays.asList(securityContext()))
-                .securitySchemes(Arrays.asList(apiKey()))
-                .select()
-                .apis(RequestHandlerSelectors.withClassAnnotation(RestController.class))
-                .paths(PathSelectors.any())
-                .build().apiInfo(apiInfo());
-    }
-
-
-    private SecurityContext securityContext() {
-        return SecurityContext.builder()
-                .securityReferences(defaultAuth())
-                .build();
-    }
-
-
-    private List<SecurityReference> defaultAuth() {
-        AuthorizationScope authorizationScope = new AuthorizationScope("global", "accessEverything");
-        AuthorizationScope[] authorizationScopes = new AuthorizationScope[1];
-        authorizationScopes[0] = authorizationScope;
-        return Arrays.asList(new SecurityReference("Authorization", authorizationScopes));
-    }
-
-
-    private ApiKey apiKey() {
-        return new ApiKey("Authorization", "Bearer", "header");
-    }
-
-
-    private ApiInfo apiInfo() {
-        String title = "SEMTLE";
-        String description = "SEMTLE API Documentation";
-        return new ApiInfoBuilder()
-                .title(title)
-                .description(description)
-                .version("1.0")
-                .build();
-    }
-}
