+++ src/main/java/com/semtleapp/semtleapp/file/exception/FileDownloadException.java
@@ -1,0 +1,11 @@
+package com.semtleapp.semtleapp.file.exception;
+
+public class FileDownloadException extends RuntimeException{
+    public FileDownloadException(String message) {
+        super(message);
+    }
+
+    public FileDownloadException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
