+++ src/main/java/com/semtleapp/semtleapp/file/exception/FileUploadException.java
@@ -1,0 +1,12 @@
+package com.semtleapp.semtleapp.file.exception;
+
+public class FileUploadException extends RuntimeException {
+    public FileUploadException(String message) {
+        super(message);
+    }
+
+    public FileUploadException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
+
