+++ src/main/java/com/semtleapp/semtleapp/file/exception/FileUploadProperties.java
@@ -1,0 +1,20 @@
+package com.semtleapp.semtleapp.file.exception;
+
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@ConfigurationProperties(prefix = "file")
+@Component
+public class FileUploadProperties {
+    //@Value("${file.upload-dir}")
+    private String uploadDir;
+
+    public String getUploadDir() {
+        return uploadDir;
+    }
+
+    public void setUploadDir(String uploadDir) {
+        this.uploadDir = uploadDir;
+    }
+}
