+++ src/main/java/com/semtleapp/semtleapp/file/service/FileService.java
@@ -1,0 +1,16 @@
+package com.semtleapp.semtleapp.file.service;
+
+import org.springframework.core.io.Resource;
+import org.springframework.http.ResponseEntity;
+
+import javax.servlet.http.HttpServletRequest;
+import java.util.Map;
+
+public interface FileService {
+
+    ResponseEntity<Resource> loadFileAsResource(HttpServletRequest request, String fileName);
+
+    ResponseEntity<Resource> showImage(Map<String, String> param);
+
+
+}
