+++ src/main/java/com/semtleapp/semtleapp/global/config/SecurityConfig.java
@@ -1,0 +1,97 @@
+package com.semtleapp.semtleapp.global.config;
+
+import com.semtleapp.semtleapp.global.jwt.JwtProvider;
+import com.semtleapp.semtleapp.global.jwt.JwtAuthenticationFilter;
+import com.semtleapp.semtleapp.global.handler.AuthenticationEntryPointHandler;
+import com.semtleapp.semtleapp.global.handler.WebAccessDeniedHandler;
+import lombok.RequiredArgsConstructor;
+import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
+import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.CorsConfigurationSource;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+
+import java.util.Arrays;
+
+@Configuration
+@RequiredArgsConstructor
+@EnableWebSecurity
+@EnableGlobalMethodSecurity(prePostEnabled = true)
+public class SecurityConfig {
+
+    private final JwtProvider jwtProvider;
+    private final AuthenticationEntryPointHandler authenticationEntryPointHandler;
+    private final WebAccessDeniedHandler webAccessDeniedHandler;
+
+    @Bean
+    public AuthenticationManager authenticationManager(AuthenticationConfiguration configuration) throws Exception {
+        return configuration.getAuthenticationManager();
+    }
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+        http.csrf().disable();
+        //http.httpBasic().disable(); // 일반적인 루트가 아닌 다른 방식으로 요청시 거절, header에 id, pw가 아닌 token(jwt)을 달고 간다. 그래서 basic이 아닌 bearer를 사용한다.
+        http.httpBasic().disable()
+                .authorizeRequests()// 요청에 대한 사용권한 체크
+                .antMatchers("/s1/**").authenticated()//url에 s1이 들어가면 jwt 검증
+                .antMatchers("/s1/m/**").hasRole("ADMIN")//회장
+                .antMatchers("/s1/as/**").hasRole("ASEM")//부회장
+                .antMatchers("/s1/bs/**").hasRole("BSEM")//총무
+                .antMatchers("/s1/cs/**").hasRole("CSEM")//서기
+                .antMatchers("/s1/ds/**").hasRole("DSEM")//홍보
+                .antMatchers("/s1/es/**").hasRole("ESEM")//스터디
+                .antMatchers("/u/**").permitAll()//여기는 jwt검증 안함
+                .and()
+                .cors()
+                .and()
+                .exceptionHandling()
+                .authenticationEntryPoint(authenticationEntryPointHandler)
+                .accessDeniedHandler(webAccessDeniedHandler)
+                .and()
+                .addFilterBefore(new JwtAuthenticationFilter(jwtProvider),
+                        UsernamePasswordAuthenticationFilter.class); // JwtAuthenticationFilter를 UsernamePasswordAuthenticationFilter 전에 넣는다
+
+        // + 토큰에 저장된 유저정보를 활용하여야 하기 때문에 CustomUserDetailService 클래스를 생성합니다.
+        http.sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS);
+
+        return http.build();
+    }
+    @Bean
+    public CorsConfigurationSource corsConfigurationSource() {
+        CorsConfiguration configuration = new CorsConfiguration();
+
+        configuration.setAllowedOriginPatterns(Arrays.asList("*"));
+        configuration.setAllowedMethods(Arrays.asList("HEAD","POST","GET","DELETE","PUT"));
+        configuration.setAllowedHeaders(Arrays.asList("*"));
+        configuration.setAllowCredentials(true);
+
+        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+        source.registerCorsConfiguration("/**", configuration);
+        return source;
+    }
+
+    @Bean
+    public WebSecurityCustomizer webSecurityCustomizer() {
+        return (web) -> web.ignoring()
+                .antMatchers("/h2-console/**")
+                .requestMatchers(PathRequest.toStaticResources().atCommonLocations());
+    }
+}
