+++ src/main/java/com/semtleapp/semtleapp/global/handler/AuthenticationCustomException.java
@@ -1,0 +1,18 @@
+package com.semtleapp.semtleapp.global.handler;
+
+import org.springframework.security.core.AuthenticationException;
+
+public class AuthenticationCustomException extends AuthenticationException {
+
+    public AuthenticationCustomException(String msg, Throwable cause) {
+        super(msg, cause);
+    }
+
+    public AuthenticationCustomException(String msg) {
+        super(msg);
+    }
+
+    public AuthenticationCustomException(ErrorCode errorCode) {
+        super(errorCode.toString());
+    }
+}
