+++ src/main/java/com/semtleapp/semtleapp/global/handler/CustomException.java
@@ -1,0 +1,10 @@
+package com.semtleapp.semtleapp.global.handler;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class CustomException extends RuntimeException{
+    ErrorCode authCode;
+}
