+++ src/main/java/com/semtleapp/semtleapp/global/handler/CustomExceptionHandler.java
@@ -1,0 +1,18 @@
+package com.semtleapp.semtleapp.global.handler;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@ControllerAdvice
+public class CustomExceptionHandler {
+
+
+    @ExceptionHandler(CustomException.class)
+    protected ResponseEntity<ErrorResponseEntity> handleCustomException(CustomException e) {
+
+        return ErrorResponseEntity.toResponseEntity(e.getAuthCode());
+    }
+
+
+}
