--- src/main/java/com/semtleapp/semtleapp/semtleuser/filter/JwtAuthenticationFilter.java
@@ -1,35 +1,0 @@
-package com.semtleapp.semtleapp.semtleuser.filter;
-
-import com.semtleapp.semtleapp.semtleuser.config.JwtProvider;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.web.filter.GenericFilterBean;
-
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
-import javax.servlet.http.HttpServletRequest;
-import java.io.IOException;
-
-@RequiredArgsConstructor
-public class JwtAuthenticationFilter extends GenericFilterBean {
-
-    private final JwtProvider jwtProvider;
-
-
-    @Override
-    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
-        //헤더에서 JWT를 받아온다.
-        String AccessToken = jwtProvider.getAccessToken((HttpServletRequest) request);
-
-        //토큰 유효성 검사
-        if(AccessToken != null && jwtProvider.validateToken(request, AccessToken)) {
-            Authentication authentication = jwtProvider.getAuthentication(AccessToken);
-            SecurityContextHolder.getContext().setAuthentication(authentication);
-        }
-
-        chain.doFilter(request, response);
-    }
-}
