--- src/main/java/com/semtleapp/semtleapp/semtleuser/handler/AuthenticationCustomException.java
@@ -1,18 +1,0 @@
-package com.semtleapp.semtleapp.semtleuser.handler;
-
-import org.springframework.security.core.AuthenticationException;
-
-public class AuthenticationCustomException extends AuthenticationException {
-
-    public AuthenticationCustomException(String msg, Throwable cause) {
-        super(msg, cause);
-    }
-
-    public AuthenticationCustomException(String msg) {
-        super(msg);
-    }
-
-    public AuthenticationCustomException(ErrorCode errorCode) {
-        super(errorCode.toString());
-    }
-}
