--- src/main/java/com/semtleapp/semtleapp/semtleuser/handler/CustomException.java
@@ -1,10 +1,0 @@
-package com.semtleapp.semtleapp.semtleuser.handler;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class CustomException extends RuntimeException{
-    ErrorCode authCode;
-}
