--- src/main/java/com/semtleapp/semtleapp/semtleuser/handler/CustomExceptionHandler.java
@@ -1,18 +1,0 @@
-package com.semtleapp.semtleapp.semtleuser.handler;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@ControllerAdvice
-public class CustomExceptionHandler {
-
-
-    @ExceptionHandler(CustomException.class)
-    protected ResponseEntity<ErrorResponseEntity> handleCustomException(CustomException e) {
-
-        return ErrorResponseEntity.toResponseEntity(e.getAuthCode());
-    }
-
-
-}
