--- src/main/java/com/semtleapp/semtleapp/semtleuser/handler/ErrorResponseEntity.java
@@ -1,24 +1,0 @@
-package com.semtleapp.semtleapp.semtleuser.handler;
-
-import lombok.Builder;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-
-@Getter
-@Builder
-public class ErrorResponseEntity {
-    private int code;
-    private String message;
-    private HttpStatus status;
-
-    public static ResponseEntity<ErrorResponseEntity> toResponseEntity(ErrorCode e) {
-        return ResponseEntity
-                .status(e.getStatus())
-                .body(ErrorResponseEntity.builder()
-                        .status(e.getStatus())
-                        .code(e.getCode())
-                        .message(e.getMessage())
-                        .build());
-    }
-}
