--- src/main/java/com/semtleapp/semtleapp/semtleuser/handler/WebAccessDeniedHandler.java
@@ -1,28 +1,0 @@
-package com.semtleapp.semtleapp.semtleuser.handler;
-
-import org.json.simple.JSONObject;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@Component
-public class WebAccessDeniedHandler implements AccessDeniedHandler {
-    @Override
-    public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException, ServletException {
-        ErrorCode errorCode = ErrorCode.ForbiddenException;
-
-        response.setContentType("application/json;charset=UTF-8");
-        response.setCharacterEncoding("utf-8");
-        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-
-        JSONObject json = new JSONObject();
-        json.put("code", errorCode.getCode());
-        json.put("message", errorCode.getMessage());
-        response.getWriter().print(json);
-    }
-}
